+++ testimonials-widget.php
@@ -3,7 +3,7 @@
  * Plugin Name: Testimonials
  * Plugin URI: http://wordpress.org/extend/plugins/testimonials-widget/
  * Description: Testimonials lets you randomly slide or list selected portfolios, quotes, reviews, or text with images or videos on your WordPress site.
- * Version: 2.15.2-alpha
+ * Version: 2.15.2-beta
  * Author: Michael Cannon
  * Author URI: http://aihr.us/about-aihrus/michael-cannon-resume/
  * License: GPLv2 or later
@@ -23,30 +23,39 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  */
-class Testimonials_Widget {
+
+if ( ! defined( 'TW_PLUGIN_DIR' ) )
+	define( 'TW_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
+
+if ( ! defined( 'TW_PLUGIN_DIR_LIB' ) )
+	define( 'TW_PLUGIN_DIR_LIB', TW_PLUGIN_DIR . '/lib' );
+
+require_once TW_PLUGIN_DIR_LIB . '/aihrus/class-aihrus-common.php';
+
+
+class Testimonials_Widget extends Aihrus_Common {
 	const ID          = 'testimonials-widget-testimonials';
 	const ITEM_NAME   = 'Testimonials';
-	const JS_KEY      = 'tw_slider_';
-	const OLD_NAME    = 'testimonialswidget';
-	const PLUGIN_FILE = 'testimonials-widget/testimonials-widget.php';
+	const PLUGIN_BASE = 'testimonials-widget/testimonials-widget.php';
 	const PT          = 'testimonials-widget';
-	const VERSION     = '2.15.2-alpha';
+	const SLUG        = 'tw_';
+	const VERSION     = '2.15.2-beta';
 
-	private static $base;
 	private static $found_posts   = 0;
 	private static $max_num_pages = 0;
 	private static $post_count    = 0;
 	private static $wp_query;
 
+	public static $class;
 	public static $cpt_category    = '';
 	public static $cpt_tags        = '';
 	public static $css             = array();
 	public static $css_called      = false;
-	public static $donate_button   = '';
 	public static $instance_number = 0;
 	public static $instance_widget = 0;
 	public static $menu_shortcodes;
 	public static $not_found       = false;
+	public static $notice_key;
 	public static $scripts         = array();
 	public static $scripts_called  = false;
 	public static $settings_link   = '';
@@ -93,6 +102,8 @@ class Testimonials_Widget {
 
 
 	public function __construct() {
+		parent::__construct();
+
 		add_action( 'admin_init', array( $this, 'admin_init' ) );
 		add_action( 'admin_menu', array( $this, 'admin_menu' ) );
 		add_action( 'init', array( $this, 'init' ) );
@@ -103,14 +114,7 @@ class Testimonials_Widget {
 
 
 	public function admin_init() {
-		self::$donate_button = <<<EOD
-<form action="https://www.paypal.com/cgi-bin/webscr" method="post" target="_top">
-<input type="hidden" name="cmd" value="_s-xclick">
-<input type="hidden" name="hosted_button_id" value="WM4F995W9LHXE">
-<input type="image" src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" name="submit" alt="PayPal - The safer, easier way to pay online!">
-<img alt="" border="0" src="https://www.paypalobjects.com/en_US/i/scr/pixel.gif" width="1" height="1">
-</form>
-EOD;
+		self::support_thumbnails();
 
 		self::$settings_link = '<a href="' . get_admin_url() . 'edit.php?post_type=' . Testimonials_Widget::PT . '&page=' . Testimonials_Widget_Settings::ID . '">' . esc_html__( 'Settings', 'testimonials-widget' ) . '</a>';
 
@@ -125,7 +129,6 @@ EOD;
 		add_filter( 'plugin_row_meta', array( $this, 'plugin_row_meta' ), 10, 2 );
 		add_filter( 'post_updated_messages', array( $this, 'post_updated_messages' ) );
 		add_filter( 'pre_get_posts', array( $this, 'pre_get_posts_author' ) );
-		self::support_thumbnails();
 	}
 
 
@@ -136,10 +139,12 @@ EOD;
 
 	public function init() {
 		add_filter( 'the_content', array( $this, 'get_single' ) );
+
 		load_plugin_textdomain( self::PT, false, 'testimonials-widget/languages' );
-		self::$base         = plugin_basename( __FILE__ );
+
 		self::$cpt_category = self::PT . '-category';
 		self::$cpt_tags     = self::PT . '-post_tag';
+
 		self::init_post_type();
 
 		$force_css_loading = tw_get_option( 'force_css_loading' );
@@ -148,8 +153,8 @@ EOD;
 	}
 
 
-	public function plugin_action_links( $links, $file ) {
-		if ( $file == self::$base )
+	public static function plugin_action_links( $links, $file ) {
+		if ( self::PLUGIN_BASE == $file )
 			array_unshift( $links, self::$settings_link );
 
 		return $links;
@@ -240,7 +245,7 @@ EOD;
 		if ( ! current_user_can( 'activate_plugins' ) )
 			return;
 
-		require_once 'lib/class-testimonials-widget-settings.php';
+		require_once TW_PLUGIN_DIR_LIB . '/class-testimonials-widget-settings.php';
 		self::init();
 		flush_rewrite_rules();
 	}
@@ -260,11 +265,10 @@ EOD;
 
 		global $wpdb;
 
-		require_once 'lib/class-testimonials-widget-settings.php';
+		require_once TW_PLUGIN_DIR_LIB . '/class-testimonials-widget-settings.php';
 		$delete_data = tw_get_option( 'delete_data', false );
 		if ( $delete_data ) {
 			delete_option( Testimonials_Widget_Settings::ID );
-			delete_option( self::OLD_NAME );
 			$wpdb->query( 'OPTIMIZE TABLE `' . $wpdb->options . '`' );
 
 			Testimonials_Widget::delete_testimonials();
@@ -307,7 +311,7 @@ EOD;
 
 
 	public static function plugin_row_meta( $input, $file ) {
-		if ( $file != self::$base )
+		if ( self::PLUGIN_BASE != $file )
 			return $input;
 
 		$disable_donate = tw_get_option( 'disable_donate' );
@@ -315,7 +319,7 @@ EOD;
 			return $input;
 
 		$links = array(
-			'<a href="http://aihr.us/about-aihrus/donate/"><img src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" alt="PayPal - The safer, easier way to pay online!" /></a>',
+			self::$donate_link,
 			'<a href="http://aihr.us/downloads/testimonials-widget-premium-wordpress-plugin/">Purchase Testimonials Premium</a>',
 		);
 
@@ -325,25 +329,17 @@ EOD;
 	}
 
 
-	public function admin_notices_2_12_0() {
-		$content  = '<div class="updated fade"><p>';
-		$content .= sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible CSS fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/23722573-Major-Changes-Since-2-10-0' ) );
-		$content .= '</p></div>';
+	public static function notice_2_12_0() {
+		$text = sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible CSS fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/23722573-Major-Changes-Since-2-10-0' ) );
 
-		echo $content;
+		parent::notice_updated( $text );
 	}
 
 
-	public function admin_notices_donate() {
+	public static function notice_donate( $disable_donate = null, $item_name = null ) {
 		$disable_donate = tw_get_option( 'disable_donate' );
-		if ( $disable_donate )
-			return;
 
-		$content  = '<div class="updated fade"><p>';
-		$content .= sprintf( esc_html__( 'Please donate $5 towards development and support of this Testimonials plugin. %s', 'testimonials-widget' ), self::$donate_button );
-		$content .= '</p></div>';
-
-		echo $content;
+		parent::notice_donate( $disable_donate, self::ITEM_NAME );
 	}
 
 
@@ -351,10 +347,10 @@ EOD;
 		$prior_version = tw_get_option( 'admin_notices' );
 		if ( $prior_version ) {
 			if ( $prior_version < '2.12.0' )
-				add_action( 'admin_notices', array( $this, 'admin_notices_2_12_0' ) );
+				self::set_notice( 'notice_2_12_0' );
 
 			if ( $prior_version < '2.15.0' )
-				add_action( 'admin_notices', array( $this, 'admin_notices_2_15_0' ) );
+				self::set_notice( 'notice_2_15_0' );
 
 			if ( $prior_version < self::VERSION )
 				do_action( 'testimonials_widget_update' );
@@ -365,103 +361,9 @@ EOD;
 		// display donate on major/minor version release
 		$donate_version = tw_get_option( 'donate_version', false );
 		if ( ! $donate_version || ( $donate_version != self::VERSION && preg_match( '#\.0$#', self::VERSION ) ) ) {
-			add_action( 'admin_notices', array( $this, 'admin_notices_donate' ) );
+			self::set_notice( 'notice_donate' );
 			tw_set_option( 'donate_version', self::VERSION );
 		}
-
-		$options = get_option( self::OLD_NAME );
-		if ( true !== $options['migrated'] )
-			$this->migrate();
-	}
-
-
-	public function migrate() {
-		global $wpdb;
-
-		$table_name       = $wpdb->prefix . self::OLD_NAME;
-		$meta_key         = '_' . self::PT . ':testimonial_id';
-		$has_table_query  = "SELECT table_name FROM information_schema.tables WHERE table_schema='{$wpdb->dbname}' AND table_name='{$table_name}'";
-		$has_table_result = $wpdb->get_col( $has_table_query );
-
-		if ( ! empty( $has_table_result ) ) {
-			// check that db table exists and has entries
-			$query = 'SELECT `testimonial_id`, `testimonial`, `author`, `source`, `tags`, `public`, `time_added`, `time_updated` FROM `' . $table_name . '`';
-
-			// ignore already imported
-			$done_import_query = 'SELECT meta_value FROM ' . $wpdb->postmeta . ' WHERE meta_key = "' . $meta_key . '"';
-			$done_import       = $wpdb->get_col( $done_import_query );
-
-			if ( ! empty( $done_import ) ) {
-				$done_import = array_unique( $done_import );
-				$query      .= ' WHERE testimonial_id NOT IN ( ' . implode( ',', $done_import ) . ' )';
-			}
-
-			$results = $wpdb->get_results( $query );
-			if ( ! empty( $results ) ) {
-				foreach ( $results as $result ) {
-					// author can contain title and company details
-					$author  = $result->author;
-					$company = false;
-
-					// ex: First Last of Company!
-					$author = str_replace( ' of ', ', ', $author );
-					// now ex: First Last, Company!
-
-					// ex: First Last, Company
-					// ex: First Last, Web Development Manager, Topcon Positioning Systems, Inc.
-					// ex: First Last, Owner, Company, LLC
-					$author     = str_replace( ' of ', ', ', $author );
-					$temp_comma = '^^^';
-					$author     = str_replace( ', LLC', $temp_comma . ' LLC', $author );
-
-					// now ex: First Last, Owner, Company^^^ LLC
-					$author = str_replace( ', Inc', $temp_comma . ' Inc', $author );
-
-					// ex: First Last, Web Development Manager, Company^^^ Inc.
-					// it's possible to have "Michael Cannon, Senior Developer" and "Senior Developer" become the company. Okay for now
-					$author = explode( ', ', $author );
-
-					if ( 1 < count( $author ) ) {
-						$company = array_pop( $author );
-						$company = str_replace( $temp_comma, ',', $company );
-					}
-
-					$author = implode( ', ', $author );
-					$author = str_replace( $temp_comma, ',', $author );
-
-					$post_data = array(
-						'post_type' => self::PT,
-						'post_status' => ( 'yes' == $result->public ) ? 'publish' : 'private',
-						'post_date' => $result->time_added,
-						'post_modified' => $result->time_updated,
-						'post_title' => $author,
-						'post_content' => $result->testimonial,
-						'tags_input' => $result->tags,
-					);
-
-					$post_id = wp_insert_post( $post_data, true );
-
-					// track/link testimonial import to new post
-					add_post_meta( $post_id, $meta_key, $result->testimonial_id );
-
-					if ( ! empty( $company ) )
-						add_post_meta( $post_id, 'testimonials-widget-company', $company );
-
-					$source = $result->source;
-					if ( ! empty( $source ) ) {
-						if ( is_email( $source ) ) {
-							add_post_meta( $post_id, 'testimonials-widget-email', $source );
-						} else {
-							add_post_meta( $post_id, 'testimonials-widget-url', $source );
-						}
-					}
-				}
-			}
-		}
-
-		$options['migrated'] = true;
-		delete_option( self::OLD_NAME );
-		add_option( self::OLD_NAME, $options );
 	}
 
 
@@ -865,7 +767,7 @@ EOF;
 
 					$autoControls = $show_start_stop ? 'autoControls: true,' : '';
 
-					$slider_var  = self::JS_KEY . $widget_number;
+					$slider_var  = self::SLUG . $widget_number;
 					$javascript .= <<<EOF
 var {$slider_var} = null;
 
@@ -1717,7 +1619,7 @@ EOF;
 
 
 	public function widgets_init() {
-		require_once 'lib/class-testimonials-widget-widget.php';
+		require_once TW_PLUGIN_DIR_LIB . '/class-testimonials-widget-widget.php';
 
 		register_widget( 'Testimonials_Widget_Widget' );
 	}
@@ -1729,7 +1631,7 @@ EOF;
 	 * @SuppressWarnings(PHPMD.UnusedLocalVariable)
 	 */
 	public function add_meta_box_testimonials_widget() {
-		require_once 'lib/class-redrokk-metabox-class.php';
+		require_once TW_PLUGIN_DIR_LIB . '/class-redrokk-metabox-class.php';
 
 		$fields = array(
 			array(
@@ -1767,7 +1669,7 @@ EOF;
 		$fields = apply_filters( 'testimonials_widget_meta_box', $fields );
 
 		$meta_box = redrokk_metabox_class::getInstance(
-			self::OLD_NAME,
+			self::ID,
 			array(
 				'title' => esc_html__( 'Testimonial Data', 'testimonials-widget' ),
 				'description' => '',
@@ -2132,12 +2034,10 @@ EOF;
 	}
 
 
-	public function admin_notices_2_15_0() {
-		$content  = '<div class="updated fade"><p>';
-		$content .= sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/28402246-Major-Change-for-2-15-0' ) );
-		$content .= '</p></div>';
+	public static function notice_2_15_0() {
+		$text = sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/28402246-Major-Change-for-2-15-0' ) );
 
-		echo $content;
+		parent::notice_updated( $text );
 	}
 
 
@@ -2203,8 +2103,9 @@ add_action( 'plugins_loaded', 'testimonialswidget_init', 99 );
 function testimonialswidget_init() {
 	require_once ABSPATH . 'wp-admin/includes/plugin.php';
 
-	if ( is_plugin_active( Testimonials_Widget::PLUGIN_FILE ) ) {
-		require_once 'lib/class-testimonials-widget-settings.php';
+	if ( is_plugin_active( Testimonials_Widget::PLUGIN_BASE ) ) {
+		require_once TW_PLUGIN_DIR_LIB . '/class-testimonials-widget-settings.php';
+
 		global $Testimonials_Widget_Settings;
 		if ( is_null( $Testimonials_Widget_Settings ) )
 			$Testimonials_Widget_Settings = new Testimonials_Widget_Settings();